rework image generation
This makes the image generation a bit more readable by dropping the use
of perl and building on dd only. Additionally a new variable ROM_SIZE is
introduced to be able to set the final flash image size for e.g.
flashrom usage. This variable can be set in the platform specific
Makefile.inc.

Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>